requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
pandas==2.2.2
openpyxl==3.1.2
